Add probe script in spec for qla2xxx WWN
The simple spec file WWN filter for qla2xxx did not discriminate between
qla2xxx FC adapters ports and tcm_fc FCoE ports. This fixes it and also checks
for target mode to be enabled (actually initiator mode to be disabled) on the
adapter before looking for ports.

The probe script that does that is also intended as a template for subsequent
ones. It lives under /var/target and outputs one port WWN per stdout line. all
extra messages and infos go to stderr.